Moved AWSMobileClient test to dedicated use cases
